+++ src/backend/javascript/tests/spec_file.rs
@@ -1,14 +1,14 @@
 #[test]
 fn simple_type_usage() {
     ::spec_to_final_compilation_unit(r#"
-@type "integer"
+@type integer("u8")
 def_native("root_test_1");
 
 def("root_test_2") => root_test_1;
 def("root_test_3") => ::root_test_1;
 
 namespace("test_ns") {
-    @type "integer"
+    @type integer("u8")
     def_native("ns_test_1");
 
     def("ns_test_2") => ::root_test_1;
